python/chatbot_cloud_util/requirements.txt
chatbot_validator

aws-xray-sdk
aws_lambda_powertools
boto3
fastjsonschema
python-dateutil
python-jose[cryptography]
redis
requests
